Update 8.md
Noticed that the heap variable was actually getting deleted so the delete in the callback wasn't necessary. Also, creating on the heap is not necessary as the stack variable is still available. I've also made a note on this in the guide since I believe a stack variable should almost always be used?